the color of the install icon now depends on whether you have the app or not
Ignore-this: 991e322a4d593c9f1d521620a98d5c33

The color is green if you don't have the app, and black if the app is already installed

darcs-hash:20141203191242-c21fe-c2eca1c16d978fa792a02fa00d847c4bc8362091

+++ src/gui/app/script_app.cpp
@@ -127,13 +127,25 @@ namespace fire
             void script_app::setup_decorations()
             {
                 INVARIANT(_app);
+                INVARIANT(_app_service);
                 REQUIRE_FALSE(_clone);
 
                 set_title(_app->name().c_str());
 
                 _clone = new QPushButton;
                 make_install(*_clone);
-                _clone->setToolTip(tr("install app"));
+
+                //color the install icon depending on whether the app is already installed
+                bool exists = _app_service->available_apps().count(_app->id());
+                if(exists) 
+                {
+                    _clone->setToolTip(tr("update app"));
+                    _clone->setStyleSheet("border: 0px; color: 'black';");
+                }
+                else
+                {
+                    _clone->setToolTip(tr("install app"));
+                }
 
                 connect(_clone, SIGNAL(clicked()), this, SLOT(clone_app()));
